peer_manager.rs renamed to tezedge_state_manager.rs + cleanup
