+++ shell/src/chain_manager.rs
@@ -45,7 +45,6 @@ use crate::mempool::mempool_prevalidator::{
 use crate::mempool::mempool_state::MempoolState;
 use crate::mempool::{CurrentMempoolStateStorageRef, MempoolPrevalidatorFactory};
 use crate::peer_branch_bootstrapper::{CleanPeerData, UpdateBranchBootstraping};
-use crate::peer_manager::ProposerHandle;
 use crate::shell_channel::{
     AllBlockOperationsReceived, BlockReceived, InjectBlock, InjectBlockOneshotResultCallback,
     ShellChannelMsg, ShellChannelRef, ShellChannelTopic,
@@ -59,6 +58,7 @@ use crate::state::synchronization_state::{
 };
 use crate::state::StateError;
 use crate::subscription::*;
+use crate::tezedge_state_manager::ProposerHandle;
 use crate::utils::dispatch_oneshot_result;
 use crate::validation;
