+++ shell/src/state/chain_state.rs
@@ -28,13 +28,13 @@ use crate::peer_branch_bootstrapper::{
     PeerBranchBootstrapper, PeerBranchBootstrapperConfiguration, PeerBranchBootstrapperRef,
     StartBranchBootstraping, UpdateBlockState, UpdateOperationsState,
 };
-use crate::peer_manager::ProposerHandle;
 use crate::shell_channel::ShellChannelRef;
 use crate::state::bootstrap_state::InnerBlockState;
 use crate::state::data_requester::{DataRequester, DataRequesterRef};
 use crate::state::head_state::CurrentHeadRef;
 use crate::state::peer_state::{DataQueuesLimits, PeerState};
 use crate::state::StateError;
+use crate::tezedge_state_manager::ProposerHandle;
 use crate::validation;
 
 /// Constants for controlling bootstrap speed
